fix(utils): synchronize grading logic between directories
- Removed duplicate `grade_dataset.py` from the `blech_clust` directory.
- Retained the correct `grade_dataset.py` in the `utils` directory with complete function definitions and examples.
- Ensured all grading functions and their usage examples remain intact and accessible in a single file location.
